# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Click",
        "version": "1.0"
    },
    "paths": {
        "/api/auction": {
            "get": {
                "tags": [
                    "2.Auction API"
                ],
                "summary": "Get App",
                "description": "Hamma auctionlarni olish",
                "operationId": "app_index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Auction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "1.Auth API"
                ],
                "summary": "Login Operator",
                "description": "Login Parol yordamida kirish Operatorlar uchun",
                "operationId": "auth_operator_login",
                "requestBody": {
                    "description": "lesson save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "contact",
                                    "password"
                                ],
                                "properties": {
                                    "contact": {
                                        "type": "string",
                                        "format": "contact",
                                        "example": "998998765678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "1.Auth API"
                ],
                "summary": "Logout Operator",
                "description": "Foydalanuvchini tizimdan chiqishi",
                "operationId": "auth_operator_logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Not found"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": []
                    },
                    "code": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            },
            "Auction": {
                "title": "Auction Title",
                "properties": {
                    "car_id ": {
                        "type": "integer"
                    },
                    "starting_price": {
                        "type": "integer"
                    },
                    "life_cycle": {
                        "type": "enum['waiting_start', 'playing', 'waiting_confirma']"
                    },
                    "start": {
                        "type": "data"
                    },
                    "finish": {
                        "type": "data"
                    },
                    "join_btn_message_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "Operator": {
                "title": "Operator Title",
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "contact": {
                        "type": "contact"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "1.Auth API",
            "description": "Authorizetsiya uchun API"
        },
        {
            "name": "2.Auction API",
            "description": "Auction uchun API"
        }
    ]
}